move duplicate code into utils.tsx
