+++ react/src/pages/analysis/Analyses.tsx
@@ -9,6 +9,7 @@ import CancelAnalysisDialog from './CancelAnalysisDialog';
 import AnalysisInfoDialog from './AnalysisInfoDialog';
 import AddAnalysisAlert from './AddAnalysisAlert';
 import SetAssigneeDialog from './SetAssigneeDialog';
+import { Analysis } from '../utils';
 
 const useStyles = makeStyles(theme => ({
     root: {
@@ -73,23 +74,6 @@ export enum PipelineStatus {
     CANCELLED = "Cancelled"
 }
 
-export interface AnalysisRow {
-    analysis_id: string;
-    pipeline_id: string; // Display pipeline name?
-    result_hpf_path: string;
-    assignee: string;  // show ID or username?
-    requester: string; // show ID or username?
-    state: PipelineStatus;
-    updated: string; // Date type maybe?
-    notes: string;
-    selected: boolean; // used for optimizing data updating
-    /*
-    More fields can go here as required;
-    MaterialTable columns only need to cover a subset of
-    fields in this interface.
-    */
-}
-
 // generate fake analysis data
 export function createAnalysis(
     analysis_id: string,
@@ -99,7 +83,7 @@ export function createAnalysis(
     requester: string,
     state: PipelineStatus,
     updated: string,
-    notes: string): AnalysisRow {
+    notes: string): Analysis {
 
     return {
         analysis_id,
@@ -139,7 +123,7 @@ type ParamTypes = {
 }
 
 // Displays notes as an interactive tooltip
-function renderNotes(rowData: AnalysisRow) {
+function renderNotes(rowData: Analysis) {
     return (
     <Tooltip title={
         <>
@@ -152,7 +136,7 @@ function renderNotes(rowData: AnalysisRow) {
 }
 
 // Returns the analysis IDs of the provided rows, optionally delimited with delim
-function rowsToString(rows: AnalysisRow[], delim?: string) {
+function rowsToString(rows: Analysis[], delim?: string) {
     let returnStr = "";
     if (delim) {
         rows.forEach((row) => returnStr = returnStr.concat(`${row.analysis_id}${delim}`));
@@ -172,10 +156,10 @@ function rowsToString(rows: AnalysisRow[], delim?: string) {
 const refreshTimeDelay = 1000 * 60;
 
 /**
- * Convert the provided JSON Array to a valid array of AnalysisRows.
+ * Convert the provided JSON Array to a valid array of Analysiss.
  */
-function jsonToAnalysisRows(data: Array<any>): AnalysisRow[] {
-    const rows: AnalysisRow[] = data.map((row, index, arr) => {
+function jsonToAnalyses(data: Array<any>): Analysis[] {
+    const rows: Analysis[] = data.map((row, index, arr) => {
         switch (row.analysis_state) {
             case 'Requested':
                 row.state = PipelineStatus.PENDING;
@@ -197,7 +181,7 @@ function jsonToAnalysisRows(data: Array<any>): AnalysisRow[] {
                 break;
         }
 
-        return { ...row, selected: false } as AnalysisRow;
+        return { ...row, selected: false } as Analysis;
     });
     return rows;
 }
@@ -205,28 +189,28 @@ function jsonToAnalysisRows(data: Array<any>): AnalysisRow[] {
 /**
  * Returns whether this analysis is allowed to be cancelled.
  */
-function cancelFilter(row: AnalysisRow) {
+function cancelFilter(row: Analysis) {
     return row.state === PipelineStatus.RUNNING || row.state === PipelineStatus.PENDING;
 }
 
 /**
  * Returns whether this analysis is allowed to be run.
  */
-function runFilter(row: AnalysisRow) {
+function runFilter(row: Analysis) {
     return row.state === PipelineStatus.ERROR || row.state === PipelineStatus.CANCELLED;
 }
 
-export default function Analysis() {
+export default function Analyses() {
     const classes = useStyles();
     const { analysis_id }= useParams<ParamTypes>();
-    const [rows, setRows] = useState<AnalysisRow[]>([]);
+    const [rows, setRows] = useState<Analysis[]>([]);
     const [detail, setDetail] = useState(false); // for detail dialog
     const [cancel, setCancel] = useState(false); // for cancel dialog
     const [direct, setDirect] = useState(false); // for add analysis dialog (re-direct)
     const [assignment, setAssignment] = useState(false); // for set assignee dialog
 
     const detailRow = analyses.find(analysis => analysis.analysis_id === analysis_id);
-    const [activeRows, setActiveRows] = useState<AnalysisRow[]>(detailRow ? [detailRow] : []);
+    const [activeRows, setActiveRows] = useState<Analysis[]>(detailRow ? [detailRow] : []);
 
     const [chipFilter, setChipFilter] = useState<string>(""); // filter by state
 
@@ -241,7 +225,7 @@ export default function Analysis() {
         fetch('/api/analyses', { method: "GET" })
         .then(response => response.json())
         .then(data => {
-            const rows = jsonToAnalysisRows(data);
+            const rows = jsonToAnalyses(data);
             setRows(rows);
         });
 
@@ -271,7 +255,7 @@ export default function Analysis() {
                         const newRows = [...rows];
                         newRows.forEach((row, index, arr) => {
                             if (row.selected && cancelFilter(row)) {
-                                const newRow: AnalysisRow = { ...newRows[index] };
+                                const newRow: Analysis = { ...newRows[index] };
                                 newRow.state = PipelineStatus.CANCELLED;
                                 newRows[index] = newRow;
                             }
@@ -310,7 +294,7 @@ export default function Analysis() {
                         const newRows = [...rows];
                         newRows.forEach((row, index, arr) => {
                             if (row.selected) {
-                                const newRow: AnalysisRow = { ...newRows[index] };
+                                const newRow: Analysis = { ...newRows[index] };
                                 newRow.assignee = username;
                                 newRows[index] = newRow;
                             }
@@ -350,7 +334,7 @@ export default function Analysis() {
 
                         if (row) {  // one row changed
                             const index = rows.indexOf(row);
-                            const newRow: AnalysisRow = { ...newRows[index] };
+                            const newRow: Analysis = { ...newRows[index] };
                             newRow.selected = !row.selected;
                             newRows[index] = newRow;
                         }
@@ -390,9 +374,9 @@ export default function Analysis() {
                             position: 'row',
                             onClick: (event, rowData) => {
                                 // We can only view details of one row at a time
-                                setActiveRows([rowData as AnalysisRow]);
+                                setActiveRows([rowData as Analysis]);
                                 setDetail(true);
-                                history.push(`/analysis/${(rowData as AnalysisRow).analysis_id}`);
+                                history.push(`/analysis/${(rowData as Analysis).analysis_id}`);
                             }
                         },
                         {
@@ -400,7 +384,7 @@ export default function Analysis() {
                             tooltip: 'Cancel analysis',
                             position: 'toolbarOnSelect',
                             onClick: (event, rowData) => {
-                                setActiveRows(rowData as AnalysisRow[]);
+                                setActiveRows(rowData as Analysis[]);
                                 setCancel(true);
                             },
                         },
@@ -416,14 +400,14 @@ export default function Analysis() {
                             tooltip: 'Run analysis',
                             position: 'toolbarOnSelect',
                             onClick: (event, rowData) => {
-                                setActiveRows(rowData as AnalysisRow[]);
+                                setActiveRows(rowData as Analysis[]);
                                 // TODO: PATCH here to start pending analyses
 
                                 // If successful...
                                 const newRows = [...rows];
                                 newRows.forEach((row, index, arr) => {
                                     if (row.selected && runFilter(row)) {
-                                        const newRow: AnalysisRow = { ...newRows[index] };
+                                        const newRow: Analysis = { ...newRows[index] };
                                         newRow.state = PipelineStatus.RUNNING;
                                         newRows[index] = newRow;
                                     }
@@ -437,7 +421,7 @@ export default function Analysis() {
                             tooltip: "Assign to...",
                             onClick: (event, rowData) => {
                                 // Data handled by SetAssigneeDialog onSubmit
-                                setActiveRows(rowData as AnalysisRow[]);
+                                setActiveRows(rowData as Analysis[]);
                                 setAssignment(true);
                             }
                         }
@@ -453,7 +437,7 @@ export default function Analysis() {
                                     return row.analysis_id === oldData?.analysis_id
                                 });
 
-                                const newRow: AnalysisRow = { ...dataUpdate[index] };
+                                const newRow: Analysis = { ...dataUpdate[index] };
 
                                 // only update the columns that are allowed
                                 newRow.notes = newData.notes;
