+++ react/src/pages/dashboard/SampleTable.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core';
 import MaterialTable from 'material-table';
 import DatasetAccordions from './DatasetAccordion';
-import { Sample } from './MockData';
+import { Sample } from '../utils';
 
 const useStyles = makeStyles(theme => ({
     table: {
